move controller to code to a file
